Resolve cinder iscsi incompatibilities
This change adds some missing config for cinder compatibility and a
couple extra tasks needed in the deployment example playbook which wire
the cinder rootwrap conf to cinder itself.

The nova container has been further extended to support the cinder
workload as well as resolve missing functionality in the nova python
enviroment.

Signed-off-by: Kevin Carter <kevin@cloudnull.com>