Create the horizon rxt image
This change adds horizon panels to the system. By default the stock horizon image
does not contain panels an operator may want to leverage. To ensure we're able to
offer a better horizon experience we're creating a new image which has some additional
panels.

A change has been made to include the new panels in horizon's configuration.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>