Add multi-stage builds to our containers
This change updates the builds for the containers to use a multistage build where
applicable. All containers will now also cleanup the pycache files when detected,
ensuring that we're building efficient images. With this change we'll now publish
container images with an epoc so that we can use point in time releases when needed.

This change adds pre-commit hooks to our setup so that we can begin running some
static analysis against our code base.

> To ensure that our tests run we've made a change to the `addons.yml` file to
  correct a syntax error.

Ansible-lint config has been added to ignore most of the infractions at this point,
we'll need to address those rule violations as we develop the tool-chain.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>