Add additional config to ensure ovn+octavia is working
This change adds some additional config to handle octavia's use of the public
URL when creating loadbalancer. While most of the calls are done over the
internalURL it seems octavia defaults to the publicURL in some cases. This
update ensures that we have a functional octavia environment in all cases.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>