Use default ssh config when connecting to bastion
During the infra-deploy playbook, ansible connects to the bastion or
launcher node. It's best that this be done with the user's default ssh
configuration in lieu of the generated one, otherwise connections to the
bastion could fail.

Use ansible.builtin.ping to check liveness which is more thorough of a
test. It will use the default ssh config, whereas wait_for will only
attempt a tcp connection, which may fail if a user has a default proxy
configured as required in their environment.