Switches prometheus to use prometheus-operator
We are not using the operator install for proemtheus. This results
in having to manually install and maintain CRDs to sync scraping
rules for exporters and many other features that comes bundled
with prometheus-operator. This PR switches the prometheus installation
to use kube-prometheus-stack helm chart which uses operator.

JIRA: OSPC-267