fix: remove unnecessary usage of mariadb root password
The mariadb root password is only used in OpenStack Helm for the db-init
and db-drop jobs. These jobs are only used when using OpenStack Helm
Infra's MariaDB and not the MariaDB operator. The operator provides its
own way to define databases, users, and create grants and that's what is
used by Genestack. The only reason that the root password is then used
is to create the secret which the db-init and db-drop jobs use. So
disable the creation of that secret and remove the need for it.

Signed-off-by: Doug Goldstein <doug.goldstein@rackspace.com>