feat: introduce gnocchi helm chart overrides
This chart begins to onboard Gnocchi, the Openstack Metrics API. For now
we are using the PostgreSQL driver for indexer storage, as it's unclear
how to change it over to MySQL. Moreover, Gnocchi recommends using
PostgreSQL anyway.

Should it be determined later that we can safely use our existing
MariaDB implementation in Genestack with no major drawbacks, we should
migrate over.

This chart is not well maintained, requires updated images, and several
patches to make it work with Genestack.