fix: update our galera cluster manifest to run the latest stable
This change is updating the mariadb-galera manifest to use the latest images
and system config. This is being done to help resolve some of the deadlock
issues being experienced when the cluster is under load.

This change also brings in a LOT of fixes into the operator, which should
help make our platform run better, generally.

All of the CRD references have been updated to use the stable release of the
operator, version 0.26.0.

> Metrics were enabled by default now that we're metric collection is on deployed.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>