feat: introduce ceilometer helm chart overrides (#128)
* feat: introduce ceilometer helm chart overrides

This begins to add the overrides for the Ceilometer helm chart.
Ceilometer provides metering, monitoring, and alarming capabilities in
Openstack for billing, performance, optimization, and capacity planning
purposes.

* fix: ceilometer-api image is deprecated

Trying to pull the wallaby image for ceilometer-api results in a 404 not
found now. ceilometer-api is in the base image as of commit
cd67930 per the upstream kolla repo, so pull that instead.

* fix: remove ceilometer-collector from config

ceilometer collector was removed from ceilometer code base[1]

[1] https://review.openstack.org/504244

* fix: ceilometer-api is fully deprecated so rm it

This disables deployment of the api pod and removes related api
configuration as ceilometer no longer has a rest API. It is simply a
worker service at this point. Gnocchi API is preferred over ceilometer.

* fix: set database keys to fake values

The database section is not used, but the base chart still tries to set
some sane default values, so to avoid confusion, just override those to
a string value that makes it obvious this section is not used.

The recommended storage location for meters and events is Gnocchi, which
is automatically discovered and used by means of keystone.

* fix: set gnocchi as the publisher

This was explicitly set to notify:// without any context as to what that
is or does. The configuration does not list that as a valid value, so
let's replace the publisher with the default, `gnocchi`.

* fix: disable the ks-endpoint job

There is no endpoint for ceilometer anymore so remove the related job
that makes a service in keystone for one.

* fix: bump ceilometer images to yoga

This was the newest tagged image that I could find for Ceilometer. We
will need to investigate building our own Ceilometer images for a later
release of Genstack.

* fix: enable db-sync to init gnocchi resource types

The helm chart has a db_sync job which executes ceilometer-upgrade which
executes the storage upgrade function that initiates the resource types
in gnocchi with their attributes.

* fix: add updated event definitions from yoga

The event definitions defined in the helm chart were very dated, update
them to match those found in the yoga release.

* fix: update gnocchi resources to yoga

The gnocchi resources were outdated. This updates them to match what was
released with Yoga.

* fix: update ceilometer meters to yoga

The existing meters were outdated. This brings them up to date with the
yoga release.

* fix: simplify pipeline sinks for now

This removes some complexity that the original helm chart introduced
which defines custom meter sinks relating to instance cpu, disk, and net
metrics. We may find ourselves disabling pollsters for individual
instances, so let's not inundate the pipeline with un-necessary
complexity yet. If we find they are useful or needed, we can re-enable
them after verifying their proper operation.

The polled metrics will still be stored in Gnocchi, just not transformed
according to the defined sinks. Iff re-introduced, these pipeline sinks
may need to be further tweaked to work with the updated event defs.

* fix: enable postgresql backup jobs

* fix: add gnocchi API replicas & enable daemonsets

This should make Gnocchi more reliable and have better overall perf.

* fix: disable resource limits for ceilometer

We don't enforce pod resource limits in other helm charts so set this to
false as the default.

* fix: remove apache2 config for ceilometer

Ceilometer no longer has a rest API so let's remove this section from
the overrides.

* fix: Add default loglevels to aid troubleshooting

When troubleshooting, it helps to raise or lower default log levels
of specific modules, setting requests related loggers to DEBUG for
example can help one diagnose ceilometer CRUD operations.

* doc: add openstack ceilometer installation

* fix: set postgresql cron backup to 0015 once a day

The default was midnight but a lot of jobs run then; kick this off a
little later to help avoid the thundering herd affect.