Adds gateway to expose internal services (#174)
This patch adds the common gateway (flex-gateway) in nginx-gateway
namespace. This gateway will be shared by all other namespaces to
expose services internally. To do so the service that is getting
exposed internally simply need to follow the fqdn naming of
*.sjc.ohthree.com and ref the flex-gateway.

Also, adding the http routes for prometheus and alertmanager that
use the above method to expose the two services internally.