Deadlock fix (#181)
* Deadlock fix

Config change to turn off retry on autocommit. This is causing database inconsistency.

* Deadlock fix also

Needed one more adjustment

* Deadlock fix 3

Further testing shows that the timeout length is not doing anything. So I removed it. The retry autocommit helps, but is not fixing completely. I needed to tell the system to rollback transactions on timeout because it was off. Testing more now, but looking better.

* This seems to do the trick