fix: update skyline to 2023.2 and ensure consistent images
This change updates our use of skyline to use 2023.2 which has several fixes
that improve qol. We've made the  init and api servers use the same image
which is better for consistency and removes potential issues between the
versions. This change removes the limits, which should let the service
perform better for longer.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>