fix: limit ingress worker processes
This change updates our worker process so that it has a base value of 4.
The worker processes used to be "auto" which would create 1 process per
core on the node. On production hardware this would result in at least
64 process and 32 threads per porc, totally 2048 threads. This change
is simply enforcing a sane limit, and documenting how we can patch a
new value into our environment when it's required.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>