feat: enable nova-ssh and tweak daemonsets (#259)
* feat: enable nova-ssh to support migrations

* fix: enable debug logging for oslo.db

OsloDB does not log very much at all, even at INFO. We want to force
debug logging so we can confirm if and how often oslodb is retrying on
DBDeadlock errors.

* fix: Speed up daemonset upgrades

We had max_unavailable set to 1 which means it can take up to n x 30+
seconds to perform rolling upgrades. That's a long time for big
environments with hundreds of nodes.