fix: set skyline-api to 2023.1 and pin SQLAlchemy
Found new installs of skyline were breaking with the latest tag. Upon
investigation, can see requests to /login are successful but further
requests to /profile and /policies result in a 401 with a response
body of `not enough values to unpack (expected 5, got 4)`. Found LP
bug https://bugs.launchpad.net/skyline-apiserver/+bug/2025755 that
seemed related.

Considering we want everything to be at 2023.1 right now, I pinned the
image version for skyline-apiserver to that, and inspected the
SQLAlchemy version to find it was 1.4.51 which is reported as a version
which yields this bug.

This has fixed the login issue in the 2023.1 image.

Co-authored-by: jorge-perez <jorge.perez@rackspace.com>