Switches glance store backend to swift
This provides the options to move the glance store
to swift. We leave the default value to pvc so as
not to break existing deployments accidently. Howerver,
providing a override on the helm command to --storage=pvc
will deploy a glance with pvc backend.

These values should be overwritten in your region
specific secrets file to ensure no mistake.