fix: rename generated secrets file and fix perms (#388)
We already have a secrets.yaml in several regions at this location in
the repository we store genestack configurations. To avoid confusion and
the possibility of that being overwritten, let's rename this newly added
secrets file.

Considering the sensitivty of the secrets therein, let's take away
global read perms too.