fix(ceilometer): update json path to vm uuid
The cinder trait for instance_id being parsed out was trying to get the
instance id from `payload.volume_attachment[0].server_id` but upon
inspection of the raw messages emitted from cinder arriving via RMQ,
it's clear that key is actually, `instance_uuid`, not `server_id`.

Now, when cinder messages traverse ceilometer, instance uuid is
populated as expected. We don't have to wait for the pollster to run.

Furthermore, explicitly define event_create and event_update on the
gnocchi volume resource to look out for volume create/attach events
resepectively, to keep the resouce up-to-date in gnocchi in near
real-time.