fix(ceilometer): process event_update msgs (#438)
* fix(ceilometer): process event_update messages

Ceilometer does not keep certain aspects about instances up-to-date.
Adding a list of events to process for event_update will ensure we apply
updated event_attributes to instance resources in gnocchi. This has the
benefit of versioning instances which is great[1].

[1] - https://gnocchi.osci.io/rest.html#history

* fix(ceilometer): separate out update events

This is necessary so that during updates we don't try to modify the
resources associated to the instance. We only want to update the
instance resource itself. There may be situations whereby we want to
update associated resources, but right now we're not trying to do that.