fix(ceilometer): tweak event_update processing (#448)
Do not use the instance_type_id field as that seems to unreliably
represent the flavor_id. From what I have observed, some messages come
through as a numerical value (the idx of the flavor in the flavors
table), and some messages come through with the uuid of the flavor.

fix(ceilometer): fix flavor_id extraction

Confirmed instnace_type_id comes through in messages as numerical index
value while instance_flavor_id is what we really want.

fix(ceilometer): use payload.node for consistency

`host` is derived from the publisher id similar to `service`, which
might be a compute api node in lieu of the actual compute hype. The node
param under payload seems sticky to the hypervisor name, so defer to
that in the gnocchi resource definition...