fix: update default_availability_zones to a value
This change updates default_availability_zones to ensure it has a
default value. While this option is supposed to assume the value
of the bound az, openstack helm returns a `<no value>` value which
makes the services very upset. By setting a value we make the service
happy, and we all want happy services.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>