fix: OSPC-612 Memcached tokens
It was reported that memcached wasn't working to the full potential
this was due to caching services not using the shared token. This
change ensures that all services will have access to the shared
token.

1. new secret created for the memcached shared token. Secret
   generation script has been updated for this new token.
2. all install scripts updated with the new shared override(s).
3. all documentation updated to point to show the updated command
   using a single source.
4. The `openstack-run-all.sh` script was removed, this was old and
   misleading.

This change will ensure that deployed clouds have better access to
memcached providing a significant performance enhancement throughout
the cloud.

Issue: https://rackspace.atlassian.net/browse/OSPC-612
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>