fix: install prometheus with helm (#531)
* fix: move prometheus chart values to helm dir

Move these helm chart yamls over to base-helm-configs. Will be
refactoring prometheus to be installed directly with helm so they make
more sense there.

Begin to make a base overlay for prometheus, it's function right now is
a no-op, just pre-plumbing for future expansion.

* fix: add helm-prometheus github workflow

This replaces the kustomize build of prometheus with the helm one
instead.

* docs: install prometheus with helm

Previously we installed prometheus without actually using helm to manage
the release. This was not ideal for a few reasons such as not being able
to use `helm rollback`, no kustomize overlay support, etc.

* fix: add prometheus deployment script and doc ref

* fix: remove wait and add positional params