feat: create a single images overlay
With this change all of the images that we use across our charts will now be defined
in a single place. This will better allow an automated pipeline to control our image
usage and ensure that we're maintaining a unified set of images for a given release,
branch, or deployment.

* All images are using the builtin kustomize image replacement solution.
* The default base image set is now defined as 2024.1
* The helm values are defining the `image-<name>` string within them, permitting
  the override to take place
* The kustomization files have been updated to all use the updated resources
  defined in the image overlay

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>