fix(rook): Tolerate the PreferNoSchedule taint (#715)
* docs(rook): Do not pin operator version in example

The latest operator version works fine at this time. Instead of pinning
it in the install example, leave a tip on how one can specify the
version if needed.

* fix(rook): Tolerate the PreferNoSchedule taint

It's common practice to have node labels such as PreferNoSchedule.
However, the osd pods do not come online unless we remove that taint
from the storage node, or tolerate it.