fix: Correct syntax errors in secret creation (#719)
* fix: Correct syntax errors in secret creation

- Missing colon after root-password for grafana-db fixed.
- Missing document start "---" fixed.

* fix: Grafana username not in b64 raises error

We still get an error if the grafana namespace is not present when
applying the secrets in bulk because it hasn't been created yet, though
that error is more clear than `illegal base64 data at input byte 4`.

Considering the user is running this with the intent to create all
needed secrets and at this time one goes into the grafana namespace,
just create that namespace as well if it doesn't already exist (for
convenience).

Note: This requires removing the explicit --namespace openstack argument
from the docs example as the grafana secret conflicts.