fix: disable multipath when it's not used
This change will disable multipath when custom_multipath is not enabled.
This is being done because ubuntu ships with multipath installed and
configured by default and while that can be helpful, it can also cause
issues with systems.

This change will ensure that our platform is using the right services
when they're enabled.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>