doc: Install Prometheus with install-chart.sh (#764)
All of the Prometheus exporters use install-chart.sh. This switches the
documentation to using install-chart.sh for Prometheus as well, which works
equivalently to install-prometheus.sh, which also gets removed here.

Also, install-chart.sh works as a generic script that it doesn't make sense to
show in-line in the documentation, so this also removes a few `--8<--` code
block markups that display the script in-line to the documentation.