docs: Use kustomize overlay for Nginx GW API (#769)
There is a patch in the `base` overlay which modifies proxy_buffers to
account for large headers which is common in our envioronment and
ultimately results in Nginx throwing a 502 and logging an error like,
"upstream sent too big header while reading response header from upstream".

This update also includes the latest changes to the nginx.conf as
obtained from a fresh install.

Starting in v1.5.0, the SnippetsFilterAPI was added which can be used in
lieu of this configuration override, though it has to be enabled
manually, and more testing is needed to implement the same.

Jira: OSPC-990
PR: #463