fix(ceilometer): Disable the ks_service job
Ceilometer does not actually need this to function properly anymore. It
used to have an API but was deprecated a while back. Creating a service
in keystone for it is mostly harmless but results in WARNING level logs
in other services when they do a service list, find ceilometer in the
output, but no related keystone endpoint exists.