FIX: HIGH PRIORITY multipath blacklist (#787)
Our current regexp pattern ^sda[0-9]* and ^sdb[0-9]* blacklists
sda1, sda2, sda3, sda4 which is what we want. Unfortunately,
it also blocks /dev/sda[a-x], /dev/sdb[a-x] which is NOT what
we want Guarav's Rally testing revealed that with a high number of
multipath devices, they naming convention starts to add the additional
alphabet letter as it scales up.

The correction: "^sd[a-b]?" per
https://www.suse.com/support/kb/doc/?id=000016563