fix(keystone): update the keystone image to fix project id hashing
There was an issue with the project id hash which had unforseen breakage
with services like octavia. While the hash length is allowed by the keystone
driver, a hard coded value in the octavia service limits project IDs to only
32 characters.

This change pulls in a new image that will resolve this hashing issue so that
new projects don't encounter this issue.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>