feat: standalone CNI (#794)
* feat: standalone CNI

This change will allow us to create some separation between what is being
done in kubespray and how we deliver our clouds. The CNI implementation is
now a standalone process, which allows us to deliver on being a fully
k8s agnostic platform.

* New scripts have been created to help facilitate the change.
* An import script has been created to help existing users convert their
  current kube-ovn deployment into one that can be managed by helm.
* Doc updates have been created to ensure that users have all of the
  information they need to succeed.
* Updates to the documentation on setup, deployment, upgrade, and
  management have been added.
* New helper scripts have been added for the setup
* Release notes for our upcoming stable release have been added
* Namespace manifests for in path services have been added

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* fix: typo old AZ referenced in ovn-setup

This change updates the ovn-setup process to use az1 when an az
is undefined. While our documented process has this defined
correctly, the service will setup a broken AZ if the AZ is not
set.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

---------

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>