fix(ceilometer): Add octavia and magnum msg urls (#850)
This adds the rabbit messaging URLs for octavia and magnum so that their
notification messages get ACKd by ceilometer. The configuration that
would process these notifications and turn them into events does not
exist, but that may come at a later date. For now, this simply vacuums
the messages out of the Queue so that they don't build up, and will
allow us to simply configure ceiloometer in the future to do something
useful with them.

This also rearranges the order of the URLs slightly which should not
impact functionality, but to me seems more logical.