chore(ceilometer): convert ceilometer helm chart from submodule to repo (#851)
* chore(gnocchi): convert gnocchi helm chart from submodule to repo

This change removes the need to carry the openstack-helm chart for the purposes of providing a ceilometer deployment.
The base helm file has been updated and simplified, reducing the values we carry to only what we need.

Related Issue: #809

* Update bin/install-ceilometer.sh

Co-authored-by: Kevin Carter <kevin@cloudnull.com>

* fix: Add missing dependencies key

Noticed db-sync pod not starting due to a bunch of missing dependencies.
None of the overridden depends were being recognized. Found the top
level dependencies key missing, added it, tested again, and confirmed
proper operation of Ceilometer in my lab in conjunction with Gnocchi.
This is good to go.

---------

Co-authored-by: Kevin Carter <kevin@cloudnull.com>
Co-authored-by: Luke Repko <luke.repko@rackspace.com>