feat(envoy): add full-featured Envoy support (#856)
This change adds Envoy support with all the basic capabilities
used today with the NGINX Gateway and additional stubs for future
optimizations.

* Documentation has been updated to separate Envoy from NGINX Gateway
  implementations
* Traffic policies have been added for basic capabilies with the ability
  to extend and create comprehensive rate limits later.

This change is a 1:1 replacement for the NGINX Gateway API ensuing it
operates with self-signed and lets encrypt certificates, supports
inplace replacements, and makes use of all the pre-defined routes/listeners.

Nice to haves added

* X-Forwarded-For and Client-IPs are now properly tracked
* Cluster policies allow for better traffic management by default
* HPA for envoy gives the system better scale

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>