fix: add iterator for IPS (#905)
This change ensures that the controller ip addresses are iterated on when discovered.
the the "fixed_ips" field from ports is a list, we should expect it to potentially
have multiple values.

This change also ensures that our comma separated list of ip addresses is properly
formated without a leading space.

The loop will now guard when creating the port so that the script is idempotent and
ensure that the list of controller IP addresses are always sorted.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>