Fix: Include alerting rules from base genestack
We are not applying the rules found in base genestack helm
config when installing prometheus. This patch changes this to
take the rules from base and also allow extra rules to be defined
in /etc/genestack directory.

This allows us to define custom rules in /etc/ while also allowing
us to define general openstack and k8s specific rule in the generl
genestack repo.

This also adds a new rules directory for this purpose.