Remove storageclassname from helm config
If not strorageClassName is sepcified the default storage class
is used. There might be environments where the `general` class
name is not used. In genestack we are now defaulting to
longhorn-general as the default. Where we need specific class
it should be an override, where we dont leave it to use the default.