fix: add random mac to all physical interfaces created
This change will ensure that our OVN setup is consistent and functional
at scale when running with multiple provider networks. This change
allows the ovn setup to generate a unique mac address per-physical
interface name using the hostname + interface name as the seed. By
setting the hostname and interface name as the seed, we'll ensure that
the mac generated is unique per-host, but consistent should the setup
tools ever be rerun.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>