feat: add per-pr smoke tests (#935)
This change will build a simple three node cloud environment on-top of
openstack flex in dfw. The job will build, evaluate, and return the
state of the build upon completion. The hyperconverged lab script is
used to run the base test.

The new job will also ensure that it cleans up resources built
no matter the state of the build.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>