fix(keystone): Update to latest keystone-rxt image
This fixes 84977a9 which adds a new image for keystone that supports
rbac. Though it also includes another bump to the image to address a few
minor changes to the rbac mapping.

Note: After this is merged, we will need to update
keystone-helm-overrides to include the new image from Quay after it has
built and uploaded successfully.