fix: Pin helm version and update keystone image (#988)
* fix(helm): Pin helm version to 3.17.3

Helm version 3.18.0 was released a few days ago and leads to (at the
very least) malformatted DB_CONNECTION strings which was noticed when
the db-sync job was failing with a traceback for keystone.

Jira: OSPC-1330

* fix(keystone): Update to latest keystone-rxt image

This fixes 84977a9 which adds a new image for keystone that supports
rbac. Though it also includes another bump to the image to address a few
minor changes to the rbac mapping.

Note: After this is merged, we will need to update
keystone-helm-overrides to include the new image from Quay after it has
built and uploaded successfully.