fix: ensure no-tenant network leakage (#1139)
The following configuration will ensure that we're not permitting LLDP
traffic over tenant networks. In the before now we were not controlling
the installation of LLDP, however, we were requiring it for QC tools.
This change ensures that we're controlling the installation of LLDP and
the config, which ensures that we're not leaking host information to
tenants of the cloud platform.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>