Fix: Ansible playbook host_setup role raid_cli_tools.yml task (#1151)
This fixes recursion error in the vars file corresponding to
the raid_cli_tools.yml task and associated problems encountered
with HP ssacli keyrings and apt installation. TESTED in IAD3
on nodes in maintenance. Previous tests did not pick up this
problem because another version of the package had already been
installed and the task was nearly wholly skipped.