feat: updated glance policy
This change will ensure that users are not able to download images
that we publish. The policy update will allow users to download
their images, but not any image that is currently public.

This change also ensures that the setting show_image_direct_url is
disabled. While this option has a default value of false and is
undefined in openstack-helm, we are making this option explicit.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>