feat: add new policy updates for octavia failover (#1167)
* feat: add new policy updates for octavia failover

This change updates the octavia policy so that failover can be handled by both an admin
and an end-user. This will empower users to be able to resolve operational issues with
loadbalancers when they pop-up.

> This was a feature request that came from spot.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* feat: updated glance policy

This change will ensure that users are not able to download images
that we publish. The policy update will allow users to download
their images, but not any image that is currently public.

This change also ensures that the setting show_image_direct_url is
disabled. While this option has a default value of false and is
undefined in openstack-helm, we are making this option explicit.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

---------

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit a8588ad071e777b4cbcdb3ee23b60f610156732d)