Remove nginx (#1261)
* chore: document exclude-from-external-load-balancers

This is documenting the wellknown label so that folks know what to add or remove
to better ensure metallb is not scheduled to specific nodes. Additionally this
label is could exist by default on nodes, which would mean it would need to be
removed to make things work.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* chore: remove NGINX-Gateway

The NGINX gateway implementation is no longer functional, this change eliminates it
from our workflows and documentation.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* fix: move nginix to removed status

---------

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
Co-authored-by: Chris Breu <chris.breu@rackspace.com>