Fix: Update cinder-volume default release/branch in Ansible playbooks
Currently: running the playbook with default values result in failure
due to cinder_release = '2024.1'. The playbook git clones the Cinder
repo and uses cinder_release in the branch checkout. It specifically
tries to checkout stable/2024.1. The upstream Cinder git repository
has moved stable/2024.1 to unmaintained/2024.1. Our default value
should be not just be bumped to 2024.2. We should instead require
declaritive selection of the full upstream Cinder branch name which
includes the Cinder release.

Proposed change specifies cinder_release_branch rather than just
cinder_version. This ensures that deployer knows that this is
absolutely going to clone the upstream repo with release branch and
pip install into the virtualenv. As an artifact or declaring the full
release branch name, the deployer will be forced to select "unmaintained"
or "stable" which should inform their decision.