feat: Make Octavia work with Kube-OVN with TLS enabled.
It works in a lab like this, but has the same rough edges as the Neutron commit,
namely, I put 'ssl:' in the install script and I believe it should choose or
install with 'tcp:' connections where we don't have Kube-OVN TLS enabled.

It felt a little hacky to put the volume mounts in
`octavia_amphora_provider.yaml.j2` in addition to the base overrides, but this
file ends up as `/tmp/octavia_helm_overrides.yaml` in a hyperconverged lab, and
the hyperconverged lab script, invoked with `-x` for extra components
(currently, k9s and _Octavia_), throws `-f /tmp/octavia_helm_overrides.yaml`
on the _Octavia_ install, and that clobbers mounting the certificate for OVN,
because it contains a list of volumes to mount to the pods, and that replaces
the list from the base overrides.

Additionally, I haven't yet verified what happens if namespace openstack
doesn't actually have a `ovn-client-tls` secret, which again amounts to the
as-of-yet not-fully-solved problem of detecting a cluster without _Kube-OVN_
TLS enabled and degrading gracefully

Also, as when doing this with Neutron, I would like to phase out using the
service IP and use service names like `ovn-nb.kube-system.svc.cluster.local`