Fix: Use default service_api policy
The service_api rule in glance policy is defined
for servcie to servcie api calls. The default policy
is to allow servcie user and admin. This rule just by itself
as it is defined removes admin user access, so this patch
removes it to set it to its default value.