fix: Make Neutron and OVN installation work without Kube-OVN TLS.
The install scripts, etc. need to account for existing environments without
Kube-OVN TLS.

Here, the install script looks at the OVN chart overrides and chooses a 'tcp'
or 'ssl' connection string based on whether we have `ENABLE_SSL` set for
Kube-OVN.

Additionally, installations without Kube-OVN TLS enabled won't have an OVN
client TLS secret in the `openstack` namespace (or the `kube-system` namespace
for that matter). Trying to make a volume from a non-existent secret would stop
the pods from starting. k8s has an option to make a volume optional, though, so
this commit also, with the secret volume in the base overrides, uses _kustomize_
to make the volume optional. I don't believe the OpenStack helm charts would
allow me to put this `optional: true` flag on the volume specification for the
pods directly, so I just patched in it with _kustomize_