fix(setup): Improve yq installation and handling
Add verbosity to the wget command so we can troubleshoot download
failures if encountered more easily, as well as very clear and
concise messaging around whether or not yq really got installed.

YQ is more necessary on the remote host than the local host that runs
the hyperconverged-lab script, so warn about it not getting installed
locally but fail if not present on the remote host as it's a
pre-requisite to running many install scripts.