fix(setup): Add retry logic to post-setup
Troubleshooting the lab build failures, post setup is being executed
within seconds after the nova install completes, which may be too fast.

If I can confirm it's a race condition, we will want a more intelligent
way to determine when cluster is ready before proceeding to post-setup.

For now, just adding dumb retry logic to the openstack commands to see
if that helps.